Remove: old code from notes, pads and meeting ended
