--- bigbluebutton-html5/imports/ui/components/chat/styles.js
@@ -1,61 +1,0 @@
-import styled from 'styled-components';
-import {
-  colorWhite,
-  colorPrimary,
-} from '/imports/ui/stylesheets/styled-components/palette';
-import { smallOnly } from '/imports/ui/stylesheets/styled-components/breakpoints';
-import { mdPaddingX } from '/imports/ui/stylesheets/styled-components/general';
-
-const Chat = styled.div`
-  background-color: ${colorWhite};
-  padding: ${mdPaddingX};
-
-  display: flex;
-  flex-grow: 1;
-  flex-direction: column;
-  justify-content: space-around;
-  overflow: hidden;
-  height: 100%;
-
-  a {
-    color: ${colorPrimary};
-    text-decoration: none;
-
-    &:focus {
-      color: ${colorPrimary};
-      text-decoration: underline;
-    }
-    &:hover {
-      filter: brightness(90%);
-      text-decoration: underline;
-    }
-    &:active {
-      filter: brightness(85%);
-      text-decoration: underline;
-    }
-    &:hover:focus{
-      filter: brightness(90%);
-      text-decoration: underline;
-    }
-    &:focus:active {
-      filter: brightness(85%);
-      text-decoration: underline;
-    }
-  }
-  u {
-    text-decoration-line: none;
-  }
-
-  ${({ isChrome }) => isChrome && `
-    transform: translateZ(0);
-  `}
-
-  @media ${smallOnly} {
-    transform: none !important;
-    &.no-padding {
-      padding: 0;
-    }
-  }
-`;
-
-export default { Chat };
