+++ bigbluebutton-html5/imports/ui/components/meeting-ended/queries.ts
