+++ bigbluebutton-html5/imports/ui/components/meeting-ended/styles.ts
