--- bigbluebutton-html5/imports/ui/components/mobile-app-modal/container.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import { withTracker } from 'meteor/react-meteor-data';
-import MobileAppModalGraphqlContainer from './mobile-app-modal-graphql/component';
-
-import MobileAppModal from './component';
-
-const MobileAppModalContainer = (props) => <MobileAppModal {...props} />;
-
-withTracker(() => {})(MobileAppModalContainer);
-
-export default MobileAppModalGraphqlContainer;
