+++ bigbluebutton-html5/imports/ui/components/notes/component.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from 'react';
 import { defineMessages, useIntl } from 'react-intl';
 import { useMutation, useSubscription } from '@apollo/client';
 import injectWbResizeEvent from '/imports/ui/components/presentation/resize-wrapper/component';
-import NotesService from '/imports/ui/components/notes/notes-graphql/service';
-import PadContainer from '/imports/ui/components/pads/container';
+import NotesService from '/imports/ui/components/notes/service';
+import PadContainer from '/imports/ui/components/pads/pads-graphql/component';
 import browserInfo from '/imports/utils/browserInfo';
 import Header from '/imports/ui/components/common/control-header/component';
 import NotesDropdown from './notes-dropdown/component';
