+++ bigbluebutton-html5/imports/ui/components/notes/hooks/useHasPermission.ts
