+++ bigbluebutton-html5/imports/ui/components/notes/notes-dropdown/service.ts
