+++ bigbluebutton-html5/imports/ui/components/notes/styles.ts
