--- bigbluebutton-html5/imports/ui/components/pads/container.jsx
@@ -1,3 +1,0 @@
-import PadContainerGraphql from './pads-graphql/component';
-
-export default PadContainerGraphql;
