--- bigbluebutton-html5/imports/ui/components/pads/sessions/container.jsx
@@ -1,16 +1,0 @@
-import { withTracker } from 'meteor/react-meteor-data';
-import Service from './service';
-
-const Container = ({ sessions }) => {
-  Service.setCookie(sessions);
-
-  return null;
-};
-
-export default withTracker(() => {
-  const sessions = Service.getSessions();
-
-  return {
-    sessions,
-  };
-})(Container);
