+++ bigbluebutton-html5/imports/ui/components/screenshare/container.jsx
@@ -17,7 +17,7 @@ import AudioService from '/imports/ui/components/audio/service';
 import MediaService from '/imports/ui/components/media/service';
 import { defineMessages } from 'react-intl';
 import { EXTERNAL_VIDEO_STOP } from '../external-video-player/mutations';
-import { PINNED_PAD_SUBSCRIPTION } from '../notes/notes-graphql/queries';
+import { PINNED_PAD_SUBSCRIPTION } from '../notes/queries';
 
 const NOTES_CONFIG = window.meetingClientSettings.public.notes;
