+++ bigbluebutton-html5/imports/ui/components/user-list/user-list-content/component.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import Styled from './styles';
 import UserListParticipants from './user-participants/user-list-participants/component';
 import ChatList from './user-messages/chat-list/component';
-import UserNotesContainer from './user-notes/container';
+import UserNotesContainer from '../user-list-graphql/user-list-content/user-notes/component';
 import TimerContainer from './timer/container';
 import GuestPanelOpenerContainer from '../user-list-graphql/user-participants-title/guest-panel-opener/component';
 import UserPollsContainer from './user-polls/container';
